feat: simplified Dockerfile from uws app
