+++ _ultimate-express/Dockerfile
@@ -1,37 +1,31 @@
 FROM node:22-slim AS base
 RUN apt update && apt install git -y
-
-FROM base AS deps
-
+ENV PNPM_HOME="/pnpm"
+ENV PATH="$PNPM_HOME:$PATH"
 RUN corepack enable
 WORKDIR /app
-# COPY pnpm-lock.yaml ./
-RUN --mount=type=cache,id=pnpm,target=/root/.local/share/pnpm/store pnpm fetch
-COPY package.json ./
-RUN --mount=type=cache,id=pnpm,target=/root/.local/share/pnpm/store pnpm install --prod
+COPY . .
 
+# Build the wasm binary
 FROM golang:1.21 AS build-go
 WORKDIR /app
 COPY . .
 RUN go mod download
-RUN GOOS=js GOARCH=wasm go build -ldflags='-s -w' -o ./bin/app.wasm main.go
+RUN GOOS=js GOARCH=wasm go build -o ./bin/app.wasm main.go
+
+FROM base AS prod-deps
+RUN --mount=type=cache,id=pnpm,target=/pnpm/store pnpm install --prod --frozen-lockfile
 
 FROM base AS build
-RUN corepack enable
+RUN --mount=type=cache,id=pnpm,target=/pnpm/store pnpm install --frozen-lockfile
+RUN pnpm run build
+
+FROM gcr.io/distroless/nodejs22-debian12:nonroot
+ENV NODE_ENV=production
 
 WORKDIR /app
-# COPY pnpm-lock.yaml ./
-RUN --mount=type=cache,id=pnpm,target=/root/.local/share/pnpm/store pnpm fetch
-COPY package.json ./
-RUN --mount=type=cache,id=pnpm,target=/root/.local/share/pnpm/store pnpm install
-COPY . .
-RUN pnpm build
-
-FROM base
-
-WORKDIR /app
-COPY --from=deps /app/node_modules /app/node_modules
-COPY --from=build /app/dist /app/dist
-COPY --from=build-go /app/bin /app/dist/bin
-ENV NODE_ENV production
-CMD ["node", "./dist/index.js"]
+COPY --from=prod-deps /app/node_modules ./node_modules
+COPY --from=build /app/dist ./
+COPY --from=build-go /app/bin ./bin
+EXPOSE 5173
+CMD ["index.js"]
