+++ _ultimate-express/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "pnpm run build:go && tsx watch index.ts",
-    "build": "tsc",
+    "build": "tsc --noCheck",
     "build:go": "GOOS=js GOARCH=wasm go build -gcflags=all='-N -l' -gcflags='-dwarflocationlists=true' -o ./bin/app.wasm ./main.go"
   },
   "keywords": [],
@@ -18,6 +18,7 @@
     "ultimate-express": "^2.0.9"
   },
   "devDependencies": {
+    "@biomejs/biome": "2.2.3",
     "@types/express": "^5.0.3",
     "@types/node": "^24.3.1",
     "tsx": "^4.20.5",
