--- _ultimate-express/types/wasm.d.ts
@@ -1,7 +1,0 @@
-declare module "*.wasm" {
-  interface WasmExports { }
-  const wasmModule: (
-    imports?: WebAssembly.Imports,
-  ) => Promise<{ instance: WebAssembly.Instance & { exports: WasmExports } }>;
-  export default wasmModule;
-}
